Add Docker image of Hubble node.
Add alpine based Dockerfile to build production-ready version of Hubble node.
Setup Docker CI pipelines for build/push and tagging images with releases.
Paralellize Node.js testing steps to decrease CI run time.
Move minimist to prod npm deps, use import instead of require.
Move chai/testing utils to seperate util file.
Add proper error catching to a number of scripts.
Update README, add Docker README. Add local dev setup instructions.
