+++ .gitignore
@@ -5,3 +5,4 @@ types/ethers-contracts/
 cache/
 artifacts/
 build/
+docker/testData
