+++ docker/.dockerignore
@@ -1,0 +1,7 @@
+artifacts/
+build/
+cache/
+node_modules/
+**/testData
+types/ethers-contracts
+genesis.json
