+++ package.json
@@ -37,12 +37,12 @@
     "@types/chai": "^4.2.18",
     "@types/chai-as-promised": "^7.1.4",
     "@types/lodash": "^4.14.170",
+    "@types/minimist": "^1.2.1",
     "@types/mocha": "^8.2.2",
     "@types/node": "^14.17.3",
     "chai": "^4.3.4",
     "chai-as-promised": "^7.1.1",
     "hardhat": "^2.3.2",
-    "minimist": "^1.2.5",
     "prettier": "^1.19.1",
     "prettier-plugin-solidity": "1.0.0-beta.11",
     "solhint": "^3.3.6",
@@ -56,6 +56,7 @@
     "ethers": "5.3.1",
     "fastify": "^3.17.0",
     "lodash": "^4.17.21",
-    "mcl-wasm": "0.4.5"
+    "mcl-wasm": "0.4.5",
+    "minimist": "^1.2.5"
   }
 }
