+++ scripts/repl.ts
@@ -11,4 +11,9 @@ async function startRepl() {
     local.context.ethers = ethers;
 }
 
-startRepl();
+startRepl()
+    .then(() => process.exit(0))
+    .catch(error => {
+        console.error(error);
+        process.exit(1);
+    });
