+++ test/slow/rollup.massMigration.test.ts
@@ -11,7 +11,8 @@ import chaiAsPromised from "chai-as-promised";
 import { getGenesisProof, MassMigrationCommitment } from "../../ts/commitments";
 import { CommonToken } from "../../ts/decimal";
 import { Result } from "../../ts/interfaces";
-import { expectRevert, hexToUint8Array, mineBlocks } from "../../ts/utils";
+import { hexToUint8Array, mineBlocks } from "../../ts/utils";
+import { expectRevert } from "../../test/utils";
 import { Group, txMassMigrationFactory } from "../../ts/factory";
 import { deployKeyless } from "../../ts/deployment/deploy";
 import { handleNewBatch } from "../../ts/client/batchHandler";
